implement actions inside the playlists
