+++ ccpp-jammming/src/components/ui/home/playlists_container/open_playlist/OpenPlaylist.jsx
@@ -3,65 +3,12 @@ import IMG from "../../../../../assets/images/ImagesHUB";
 import usePlaylistInfo from "../../../../../hooks/user_hooks/usePlaylistInfo"
 import PlaylistTrack from "./track/PlaylistTrack";
 import usePlaylistActions from "../../../../../hooks/user_hooks/usePlaylistActions";
+import useUserInfo from "../../../../../hooks/user_hooks/useUserInfo";
 
 function OpenPlaylist({ playlistData, onBackClick, onPlayButton, onArtistClick, onAlbumClick, playTrack, pauseTrack, accessToken }) {
     const { playlistTracksArr } = usePlaylistInfo({ playlistData, accessToken });
-    const { editPlaylistName, unfollowPlaylist } = usePlaylistActions({ accessToken });
-
-    const handleEditPlaylist = async () => {
-        const newPlaylistName = prompt("Enter the new playlist name:", playlistData.playlistTitle);
-        if (newPlaylistName && newPlaylistName.trim() !== "") {
-            try {
-                // Call the API to update the playlist name
-                await editPlaylistName(playlistData.playlistId, newPlaylistName);
-
-                // Immediately update the playlist name locally in userPlaylistsArr
-                setUserPlaylistsArr((prevPlaylists) =>
-                    prevPlaylists.map((playlist) =>
-                        playlist.playlistId === playlistData.playlistId
-                            ? { ...playlist, playlistTitle: newPlaylistName }
-                            : playlist
-                    )
-                );
-
-                await refetchPlaylists(newPlaylistName, playlistData.playlistId);
-
-            } catch (error) {
-                console.error("Error updating playlist or re-fetching:", error);
-            }
-        }
-    };
-
-    // Function to handle sharing the playlist URL
-    const handleSharePlaylist = () => {
-        const playlistUrl = `https://open.spotify.com/playlist/${playlistData.playlistId}`;
-        navigator.clipboard.writeText(playlistUrl)
-            .then(() => {
-                alert(`Playlist URL copied to clipboard!`);
-            })
-            .catch((error) => {
-                console.error("Error copying playlist URL:", error);
-            });
-    };
-
-    const handleUnfollowPlaylist = async () => {
-        const confirmUnfollow = window.confirm("Are you sure you want to unfollow this playlist?");
-        if (confirmUnfollow) {
-            try {
-                await unfollowPlaylist(playlistData.playlistId);
-
-                // Remove the unfollowed playlist locally
-                setUserPlaylistsArr((prevPlaylists) =>
-                    prevPlaylists.filter(playlist => playlist.playlistId !== playlistData.playlistId)
-                );
-
-                await refetchPlaylists(); 
-
-            } catch (error) {
-                console.error("Error unfollowing playlist:", error);
-            }
-        }
-    };
+    const { setUserPlaylistsArr, refetchPlaylists, editPlaylists } = useUserInfo({accessToken});
+    const { handleEditPlaylist, handleSharePlaylist, handleUnfollowPlaylist, newEditedName } = usePlaylistActions({ playlistData, editPlaylists, refetchPlaylists, setUserPlaylistsArr, accessToken });
 
     return (
         <>
@@ -73,13 +20,13 @@ function OpenPlaylist({ playlistData, onBackClick, onPlayButton, onArtistClick, 
                         </a>
                     </div>
                     <div id="title-col" className="col d-flex flex-column justify-content-center align-items-start">
-                        <h3 className="align-items-center">{playlistData.playlistTitle}</h3>
+                        <h3 className="align-items-center">{newEditedName}</h3>
                     </div>
                     <div id="checkmark-col" className="col-auto d-flex flex-column justify-content-center align-items-center">
                         <img id="saved-icon" src={IMG.savedPNG} alt="saved icon" width="27px"/>
                     </div>
                     <div id="edit-button-col" className="col-auto d-flex flex-column justify-content-center align-items-center">
-                        <a id="edit-button" type="button" onClick={handleSharePlaylist}>
+                        <a id="edit-button" type="button" onClick={handleEditPlaylist}>
                             <img src={IMG.pencilPNG} alt="saved icon" width="27px"/>
                         </a>
                     </div>
@@ -89,8 +36,13 @@ function OpenPlaylist({ playlistData, onBackClick, onPlayButton, onArtistClick, 
                         </a>
                     </div>
                     <div id="delete-button-col" className="col-auto d-flex flex-column justify-content-center align-items-center">
-                        <a id="delete-button" type="button" onClick={handleUnfollowPlaylist}>
-                            <img src={IMG.trashBinPNG} alt="saved icon" width="27px"/>
+                        <a id="delete-button" type="button" onClick={() => {
+                                handleUnfollowPlaylist();
+                                setTimeout(() => {
+                                    onBackClick();
+                                }, 500); 
+                            }}>
+                            <img src={IMG.trashBinPNG} alt="saved icon" width="27px" />
                         </a>
                     </div>
                 </header>
