+++ ccpp-jammming/src/hooks/user_hooks/usePlaylistActions.jsx
@@ -1,7 +1,8 @@
 import { useState } from "react";
 import axios from "axios";
 
-function usePlaylistActions({ accessToken }) {
+function usePlaylistActions({ playlistData, editPlaylists, refetchPlaylists, setUserPlaylistsArr, accessToken }) {
+    const [newEditedName, setNewEditedName] = useState(playlistData.playlistTitle)
     
     // Function to edit an existing playlist name
     const editPlaylistName = async (playlistId, newPlaylistName) => {
@@ -45,7 +46,66 @@ function usePlaylistActions({ accessToken }) {
         }
     };
 
-    return { editPlaylistName, unfollowPlaylist };
+    //--------------------------------------------------------------------------------------------------------------
+
+    const handleEditPlaylist = async () => {
+        const newPlaylistName = prompt("Enter the new playlist name:", playlistData.playlistTitle);
+        if (newPlaylistName && newPlaylistName.trim() !== "") {
+            try {
+                // Call the API to update the playlist name
+                await editPlaylistName(playlistData.playlistId, newPlaylistName);
+
+                // Immediately update the playlist name locally in userPlaylistsArr
+                setUserPlaylistsArr((prevPlaylists) =>
+                    prevPlaylists.map((playlist) =>
+                        playlist.playlistId === playlistData.playlistId
+                            ? { ...playlist, playlistTitle: newPlaylistName }
+                            : playlist
+                    )
+                );
+
+                await editPlaylists(newPlaylistName, playlistData.playlistId);
+
+                setNewEditedName(newPlaylistName)
+
+            } catch (error) {
+                console.error("Error updating playlist or re-fetching:", error);
+            }
+        }
+    };
+
+    // Function to handle sharing the playlist URL
+    const handleSharePlaylist = () => {
+        const playlistUrl = `https://open.spotify.com/playlist/${playlistData.playlistId}`;
+        navigator.clipboard.writeText(playlistUrl)
+            .then(() => {
+                alert(`Playlist URL copied to clipboard!`);
+            })
+            .catch((error) => {
+                console.error("Error copying playlist URL:", error);
+            });
+    };
+
+    const handleUnfollowPlaylist = async () => {
+        const confirmUnfollow = window.confirm("Are you sure you want to unfollow this playlist?");
+        if (confirmUnfollow) {
+            try {
+                await unfollowPlaylist(playlistData.playlistId);
+
+                // Remove the unfollowed playlist locally
+                setUserPlaylistsArr((prevPlaylists) =>
+                    prevPlaylists.filter(playlist => playlist.playlistId !== playlistData.playlistId)
+                );
+
+                await refetchPlaylists(); 
+
+            } catch (error) {
+                console.error("Error unfollowing playlist:", error);
+            }
+        }
+    };
+
+    return { handleEditPlaylist, handleSharePlaylist, handleUnfollowPlaylist, newEditedName };
 }
 
 export default usePlaylistActions;
